Eliminate PL_OP_SLAB_ALLOC
This commit eliminates the old slab allocator.  It had bugs in it, in
that ops would not be cleaned up properly after syntax errors.  So why
not fix it?  Well, the new slab allocator *is* the old one fixed.

Now that this is gone, we don’t have to worry as much about ops leak-
ing when errors occur, because it won’t happen any more.

Recent commits eliminated the only reason to hang on to it:
 PERL_DEBUG_READONLY_OPS required it.