todo.pod: Clean up entries related to op slabs
Note about storing the current pad with the ops:  If each slab belongs
to a CV (which is the case), then we would only ever walk the slab to
free its ops when the CV itself is freed.  When that happens, the pad
is about to freed anyway, so freeing pad entries for each op is unnec-
essary, and has probably always been so.  Note that cv_undef actually
sets PL_comppad and PL_curpad to null before freeing ops, to avoid
having to have them point at the right pad, and probably also to avoid
the unnecessary overhead of tracking which pad entries are available
for reuse.