Time::HiRes: fix compiler warnings
On some platforms the 1st arg to get/setitimer is an enum, which
causes -Wc++-compat to complain about passing an int instead.
Since we don't portably know the enum type (on my linux system
it's __itimer_which_t), we cant reliably cast it. So just temporarily
disable -Wc++-compat.