add new API routine sv_get_backrefs()
sv_get_backrefs() encapsulates the logic required to determine if
a scalar is a referent of a weakref, and return the weakref(s) that
reference it.

This replaces code that uses hv_backreferences_p() and also code that
use mg_find() and PERL_MAGIC_backref. Isolating XS authors from
the complexity involved. See perl #123473 for details related to this
patch.