Socket.xs: avoid integer overflow
As spotted by 'clang -fsanitize=undefined',
the shifting of the bytes to form an IPv4 address results in an
intermediate signed int, which clang detects as overflowing when <<24.
So ensure all the values to be shifted are unsigned.

Typical clang error:

Socket.xs:796:30: runtime error: left shift of 193 by 24 places cannot be represented in type 'int'