Compress-Raw-Zlib: avoid undefined warning
With 'clang -fsanitize=undefined', when f==-1, ((f)<<1) gives this warning:

deflate.c:886:54: runtime error: left shift of negative value -1

Fix it by doing ((f)*2) instead. Any half-decent compiler should spot the
'*2' and implement it efficiently underneath as a shift or add or
whatever.

NB: this commit modifies a file which is part of a separate distribution,
libbzip2, so may have licensing implications.